refactored out draw control functions to PanelUtils.h
